+++ CHANGELOG.md
@@ -10,6 +10,8 @@ that you can set version constraints properly.
 
 #### [Unreleased][unreleased] -
 
+* Added `expandWidth` method
+
 #### [v2.0.2][v2.0.2] - 2017-11-21
 
 * Collapse macOS build config to eliminate some duplication and typos